+++ src/main/java/com/bizcub/simpleDatapacks/mixin/FeatureFlagsMixin.java
@@ -1,9 +1,11 @@
 package com.bizcub.simpleDatapacks.mixin;
 
+import org.spongepowered.asm.mixin.Mixin;
+
+//? >=1.19.3 {
 import com.bizcub.simpleDatapacks.Main;
 import net.minecraft.world.flag.FeatureFlag;
 import net.minecraft.world.flag.FeatureFlagRegistry;
-import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -18,3 +20,9 @@ public class FeatureFlagsMixin {
         }
     }
 }
+
+//?} else {
+/*@Mixin(value = {})
+public class FeatureFlagsMixin {
+
+}*///?}