+++ src/main/java/com/bizcub/simpleDatapacks/mixin/ServerPacksSourceMixin.java
@@ -1,9 +1,11 @@
 package com.bizcub.simpleDatapacks.mixin;
 
+import org.spongepowered.asm.mixin.Mixin;
+
+//? >=1.19.3 {
 import com.bizcub.simpleDatapacks.Main;
 import net.minecraft.server.packs.repository.Pack;
 import net.minecraft.server.packs.repository.ServerPacksSource;
-import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -16,3 +18,9 @@ public class ServerPacksSourceMixin {
         if (!Main.getConfig().enableFeatures()) cir.cancel();
     }
 }
+
+//?} else {
+/*@Mixin(value = {})
+public class ServerPacksSourceMixin {
+
+}*///?}